test(netcdf): split the 2953-line test_plot.py by concern (Theme E.1)
Break the 26-class plot file into six concern modules — test_plot_{selectors,coords,render,
facet,animate,lazy}.py — and extract the nine shared in-memory NetCDF builders + render
side-effects into tests/netcdf/_plot_helpers.py. Each concern file re-declares the cleopatra
importorskip guard (Config + Agg backend always; ArrayGlyph/_FacetGrid where used) and imports
the helpers/make_plot_3d_nc it needs. Pure move: 147 tests collected before and after, all
green under Agg. No test logic changes.
